+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # pending
 
+  * Fix a bug with request/response cause message to end up on the error queue incorrectly 
   * Change delivery to have a list of delivery listeners, so we can hang things a delivery happening.
 
 # 0.1.14